requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Environment
python-dotenv==1.0.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Browser Automation
playwright==1.41.0
